moved header files to include. implemented smart pointer in main. added header comments. minor refactoring.
